Polaroid Collectables, Game Saving, Menu Updates, All levels added to build index
Made the script for the polaroid. Once collected it will be saved and viewable from each level and allow the player to start that level in the menu.

Game is now saved when picking up polaroid.

Updated menu with a start screen before the actual menu to allow the artist to show their work without buttons infront of it.

Created a scene fore each stage and put them in the main folder.